# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hi-tech Multimedia API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  name: string;
  phone: string;
  email: string;
  eventType: string;
  eventDate: string;
  /** @nullable */
  message?: string | null;
  status: BookingStatus;
  createdAt: string;
}

export interface BookingInput {
  name: string;
  phone: string;
  email: string;
  eventType: string;
  eventDate: string;
  message?: string;
}

export type BookingUpdateStatus = typeof BookingUpdateStatus[keyof typeof BookingUpdateStatus];


export const BookingUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  name?: string;
  phone?: string;
  email?: string;
  eventType?: string;
  eventDate?: string;
  message?: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface GalleryItem {
  id: number;
  title: string;
  category: string;
  imageUrl: string;
  /** @nullable */
  description?: string | null;
  selected: boolean;
  featured: boolean;
}

export interface GalleryItemInput {
  title: string;
  category: string;
  imageUrl: string;
  description?: string;
}

export interface GalleryItemUpdate {
  title?: string;
  category?: string;
  imageUrl?: string;
  description?: string;
  selected?: boolean;
  featured?: boolean;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  cancelled: 'cancelled',
} as const;

export interface InvoiceItem {
  description: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  clientName: string;
  clientEmail: string;
  /** @nullable */
  clientPhone?: string | null;
  /** @nullable */
  eventType?: string | null;
  /** @nullable */
  eventDate?: string | null;
  items: InvoiceItem[];
  subtotal: number;
  tax: number;
  total: number;
  /** @nullable */
  notes?: string | null;
  status: InvoiceStatus;
  createdAt: string;
}

export interface InvoiceInput {
  clientName: string;
  clientEmail: string;
  clientPhone?: string;
  eventType?: string;
  eventDate?: string;
  items: InvoiceItem[];
  tax?: number;
  notes?: string;
}

export interface Stats {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  totalRevenue: number;
  totalInvoices: number;
}