# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hi-tech Multimedia API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bookings
    description: Booking operations
  - name: services
    description: Services operations
  - name: gallery
    description: Gallery operations
  - name: invoices
    description: Invoice operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery items
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Add a gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemInput"
      responses:
        "201":
          description: Gallery item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"

  /gallery/{id}:
    patch:
      operationId: updateGalleryItem
      tags: [gallery]
      summary: Update a gallery item (toggle selected/featured)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemUpdate"
      responses:
        "200":
          description: Gallery item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
        "404":
          description: Not found
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found

  /stats:
    get:
      operationId: getStats
      tags: [bookings]
      summary: Get dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Booking:
      type: object
      required: [id, name, phone, email, eventType, eventDate, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        createdAt:
          type: string

    BookingInput:
      type: object
      required: [name, phone, email, eventType, eventDate]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        message:
          type: string

    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        message:
          type: string

    Service:
      type: object
      required: [id, name, description, price, category]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]

    GalleryItem:
      type: object
      required: [id, title, category, imageUrl, selected, featured]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        description:
          type: ["string", "null"]
        selected:
          type: boolean
        featured:
          type: boolean

    GalleryItemInput:
      type: object
      required: [title, category, imageUrl]
      properties:
        title:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        description:
          type: string

    GalleryItemUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        description:
          type: string
        selected:
          type: boolean
        featured:
          type: boolean

    Invoice:
      type: object
      required: [id, invoiceNumber, clientName, clientEmail, items, subtotal, tax, total, status, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: ["string", "null"]
        eventType:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, paid, cancelled]
        createdAt:
          type: string

    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice, total]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number

    InvoiceInput:
      type: object
      required: [clientName, clientEmail, items]
      properties:
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        tax:
          type: number
        notes:
          type: string

    Stats:
      type: object
      required: [totalBookings, pendingBookings, confirmedBookings, totalRevenue, totalInvoices]
      properties:
        totalBookings:
          type: integer
        pendingBookings:
          type: integer
        confirmedBookings:
          type: integer
        totalRevenue:
          type: number
        totalInvoices:
          type: integer