# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hi-tech Multimedia API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  name: string;
  phone: string;
  email: string;
  eventType: string;
  eventDate: string;
  /** @nullable */
  message?: string | null;
  status: BookingStatus;
  createdAt: string;
}