# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hi-tech Multimedia API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingUpdateStatus } from './bookingUpdateStatus';

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  name?: string;
  phone?: string;
  email?: string;
  eventType?: string;
  eventDate?: string;
  message?: string;
}