# lib/api-zod/src/generated/types/galleryItemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hi-tech Multimedia API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItemInput {
  title: string;
  category: string;
  imageUrl: string;
  description?: string;
}