# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hi-tech Multimedia API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}